根据Spring Quartz实现热帖排行功能 算法：log(精华分 + 评论数*10 + 点赞数*2 + 收藏数*2) + (发布时间 – 元纪元)

+++ src/main/java/com/nowcoder/community/entity/DiscussPost.java
@@ -104,7 +104,7 @@ public class DiscussPost {
         return score;
     }
 
-    public void setScore(int score) {
+    public void setScore(double score) {
         this.score = score;
     }
